refactor: split gateway access and response helpers
